gobject: Add property dispatch notification
The newly added API allows calling a function when the notification
queue is thawed. Up until now, the only way to achieve this
functionality was to subclass a GObject type and override the
GObjectClass.dispatch_property_notify virtual function. This is
obviously suboptimal, as it prevents getting bulk notification of
properties for any generic GObject instance.

https://bugzilla.gnome.org/show_bug.cgi?id=617446