Modify the 'construct' property semantics
Construct properties come in various kinds:

  • properties that may be passed to the constructor
  • properties that may only be passed to the constructor
  • properties that must be passed to the constructor

GParamSpec only covers the first two cases, using the CONSTRUCT and the
CONSTRUCT_ONLY GParamFlags.

GProperty behaves the same way, but the CONSTRUCT GPropertyFlags value
only works in conjuction with other flags:

  • if the property is has both the CONSTRUCT and WRITABLE bits set
    then the property may be passed to the constructor and can also
    be set after the instance has been constructed; this behaviour
    matches the CONSTRUCT GParamFlags.
  • if the property only has the CONSTRUCT bit set then the property
    may only be used in the construction sequence; this behaviour
    matches the CONSTRUCT_ONLY GParamFlags.

When we set the GPropertyFlags and convert them into GParamFlags we
check for the G_PROPERTY_FLAGS_WRITABLE in order to decide which
GParamFlags value, between G_PARAM_CONSTRUCT and G_PARAM_CONSTRUCT_ONLY,
has to be set on the GParamSpec class.